+++ libs/api/domains/driving-license/src/lib/util/constants.ts
@@ -2,3 +2,5 @@ export const DAYS = 24 * 3600 * 1000
 
 // TODO: max age is 6 months - eventually this should count for days of the month
 export const DRIVING_ASSESSMENT_MAX_AGE = 182 * DAYS
+
+export const LICENSE_RESPONSE_API_VERSION = '2.0'
