+++ libs/clients/driving-license/README.md
@@ -10,6 +10,42 @@ The client is generated from a copy of the openApi document
 yarn nx run clients-driving-license:schemas/openapi-generator
 ```
 
+### Import into other NestJS modules
+
+Add the service to your module imports:
+
+```typescript
+import { DrivingLicenseApiModule } from '@island.is/clients/driving-license'
+
+@Module({
+  imports: [
+    DrivingLicenseApiModule.register({
+      xroadBaseUrl: XROAD_BASE_URL,
+      xroadPath: XROAD_PATH,
+      xroadClientId: XROAD_CLIENT_ID,
+      secret: DRIVING_LICENSE_SECRET,
+      replaceInPath: {
+        from: DRIVING_LICENSE_PATH_REPLACEMENT_FROM,
+        to: DRIVING_LICENSE_PATH_REPLACEMENT_TO,
+      }
+    }),
+  ],
+})
+```
+
+Then you'll have access to VMST APIs:
+
+```typescript
+import { OkuskirteiniApi } from '@island.is/clients/vmst'
+
+@Injectable()
+export class SomeService {
+  constructor(private drivingLicenseApi: OkuskirteiniApi) {}
+
+  // etc...
+}
+```
+
 ## Code owners and maintainers
 
 - [Kosmos & Kaos](https://github.com/orgs/island-is/teams/kosmos-og-kaos/members)
