+++ libs/clients/driving-license/src/index.ts
@@ -1,1 +1,5 @@
+export {
+  DrivingLicenseApiModule,
+  DrivingLicenseConfig,
+} from './lib/drivingLicense.module'
 export * from '../gen/fetch'
