+++ libs/clients/driving-license/src/lib/drivingLicense.module.ts
@@ -1,0 +1,73 @@
+import { DynamicModule } from '@nestjs/common'
+import fetch from 'isomorphic-fetch'
+import { createWrappedFetchWithLogging } from './utils'
+import { isRunningOnEnvironment } from '@island.is/shared/utils'
+import {
+  Configuration,
+  FetchParams,
+  Middleware,
+  OkuskirteiniApi,
+  RequestContext,
+} from '../../gen/fetch'
+
+interface PathReplacement {
+  from: string
+  to: string
+}
+
+export interface DrivingLicenseConfig {
+  xroadBaseUrl: string
+  xroadPath: string
+  xroadClientId: string
+  secret: string
+  replaceInPath?: PathReplacement
+}
+
+// This is maybe a bit of a hacky fix for the API being different in production
+// than it is on dev - apparently this is due for an update
+const replacePathMiddleware = ({ from, to }: PathReplacement) => {
+  return {
+    pre: async (context: RequestContext) => {
+      return {
+        init: context.init,
+        url: context.url.replace(from, to),
+      } as FetchParams
+    },
+  } as Middleware
+}
+
+const isRunningOnProduction = isRunningOnEnvironment('production')
+
+export class DrivingLicenseApiModule {
+  static register(config: DrivingLicenseConfig): DynamicModule {
+    return {
+      module: DrivingLicenseApiModule,
+      providers: [
+        {
+          provide: OkuskirteiniApi,
+          useFactory: () =>
+            new OkuskirteiniApi(
+              new Configuration({
+                ...(config.replaceInPath
+                  ? {
+                      middleware: [replacePathMiddleware(config.replaceInPath)],
+                    }
+                  : {}),
+                fetchApi: isRunningOnProduction
+                  ? fetch
+                  : createWrappedFetchWithLogging,
+                headers: {
+                  'X-Road-Client': config.xroadClientId,
+                  SECRET: config.secret,
+                  'Content-type': 'application/json',
+                  Accept: 'application/json',
+                },
+                basePath: `${config.xroadBaseUrl}/${config.xroadPath}`,
+              }),
+            ),
+        },
+      ],
+      exports: [OkuskirteiniApi],
+    }
+  }
+}
