+++ workspace.json
@@ -938,14 +938,6 @@
             "passWithNoTests": true
           },
           "outputs": ["coverage/libs/api/domains/driving-license"]
-        },
-        "schemas/external-openapi-generator": {
-          "builder": "@nrwl/workspace:run-commands",
-          "options": {
-            "commands": [
-              "nx run clients-driving-license:schemas/openapi-generator"
-            ]
-          }
         }
       }
     },
@@ -2087,14 +2079,6 @@
           "options": {
             "command": "graphql-codegen --config libs/application/templates/driving-assessment-approval/codegen.yml"
           }
-        },
-        "schemas/external-openapi-generator": {
-          "builder": "@nrwl/workspace:run-commands",
-          "options": {
-            "commands": [
-              "nx run clients-driving-license:schemas/openapi-generator"
-            ]
-          }
         }
       }
     },
@@ -2172,14 +2156,6 @@
           "builder": "@nrwl/workspace:run-commands",
           "options": {
             "command": "graphql-codegen --config libs/application/templates/driving-license/codegen.yml"
-          }
-        },
-        "schemas/external-openapi-generator": {
-          "builder": "@nrwl/workspace:run-commands",
-          "options": {
-            "commands": [
-              "nx run clients-driving-license:schemas/openapi-generator"
-            ]
           }
         }
       }
@@ -3129,10 +3105,10 @@
             "passWithNoTests": true
           }
         },
-        "schemas/openapi-generator": {
+        "schemas/external-openapi-generator": {
           "builder": "@nrwl/workspace:run-commands",
           "options": {
-            "command": "yarn openapi-generator -o libs/clients/driving-license/gen/fetch -i libs/clients/driving-license/src/openapi_doc.json"
+            "command": "yarn openapi-generator -o libs/clients/driving-license/gen/fetch -i libs/clients/driving-license/src/clientConfig.json"
           }
         }
       }
